FEM: import tools, move stats calculation into result tools
